+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Post v1 a change is further considered breaking if the state file format is inco
 
 ### Upcoming
 
+*Minor*
+
+* Updated minimum and maximum number of roles with `--random-extra-roles`.
+
 ### v1.1.0.0
 
 *Major*